Brings 114 commits of largesim-m1 (and main through it) onto the Rhea
flavor-mixing branch. Nine files conflicted; the resolutions:

Keep-both, no semantic choice: the ENABLE_TORCH link block next to the
retargeted two_punctures PROBLEM match (CMakeLists.txt), the gravity
include (driver.cpp), the new *_parabolic_stagen task lists next to our
rewrapped opsplit line (meshblock_pack.cpp), our Rhea single-zone pgen
declarations next to the gravity/unit-test ones (pgen.hpp), both of the
independently-added FlavorMixType and RadiationM1ClosureSolver enums
(radiation_m1_params.hpp), and all four non-thermal eta/abs clamps
(radiation_m1_calc_opacities_nurates.cpp, rewrapped to 90 cols).

Took upstream's AMR boundary fix (e1605bbb, #748) wholesale, since our
side of all three of its conflicts was line-rewrapping only:

  - radiation_m1_bcs.cpp: RadiationM1BCs split into a fine-array pass
    (after prolongation) plus a new RadiationM1BCsCoarse called from
    Prolongate, replacing the single multilevel-aware entry point.
  - radiation_m1_tasks.cpp: M1 chain is now recvu -> prol -> bcs ->
    newdt. Our M1_flvmix task keeps its place between rkupdt and restu.
  - dyn_grmhd.cpp: same Prolongate/ApplyPhysicalBCs swap in the MHD
    AfterTimeIntegrator queue, plus upstream's ctor member-init order,
    which is the one matching the declaration order in dyn_grmhd.hpp.

kokkos: kept ours (4.7.04), which already contains upstream's 4.7.02, so
their 4.4.00 -> 4.7.02 bump is subsumed rather than lost.

Verified with a from-scratch CUDA + LibTorch build (SERIAL+CUDA,
NURATES=On, TORCH=On): 217 TUs, zero errors, and no warnings in
radiation_m1 or the Rhea files. This build also requires bns_nurates at
d48bfc5 or newer -- upstream's non-thermal rate separation reads the
split number coefficients (eta_0_th, eta_0_non_th, kappa_0_a_th,
kappa_0_a_non_th) that older checkouts do not define.

Note for existing Rhea runs: upstream changed defaults that none of our
Rhea athinputs pin -- rad_N_floor 1e-77 -> 1e-27 (harmless here, our N
is ~1e-6), and for nurates use_dm_eff true -> false, use_decay false ->
true, plus a new eq_warmup_cycles = 1. The nurates changes will move the
TOV Rhea results, so runs/tov_m1_rhea_aurora/ is no longer a comparable
baseline.

+++ src/outputs/coarsened_binary.cpp
@@ -296,9 +296,6 @@ void CoarsenedBinaryOutput::LoadOutputData(Mesh *pm) {
 //   All MeshBlocks are written to the same file.
 
 void CoarsenedBinaryOutput::WriteOutputFile(Mesh *pm, ParameterInput *pin) {
-  // check if slicing
-  //bool bin_slice = (out_params.slice1 || out_params.slice2 || out_params.slice3);
-
   // create filename: "cbin_"+"file_id"+"_"+"coarsening_factor"+"/file_basename"
   // + "." + "file_id" + "." + XXXXX + ".cbin"
   // where XXXXX = 5-digit file_number
@@ -353,14 +350,14 @@ void CoarsenedBinaryOutput::WriteOutputFile(Mesh *pm, ParameterInput *pin) {
       << "  variables:  ";
   if (out_params.compute_moments) {
     // need to write the label for each of the 4 moments
-    for (int n=0; n<outvars.size(); n++) {
+    for (size_t n=0; n<outvars.size(); n++) {
       msg << outvars[n].label.c_str() << "_1st  ";
       msg << outvars[n].label.c_str() << "_2nd  ";
       msg << outvars[n].label.c_str() << "_3rd  ";
       msg << outvars[n].label.c_str() << "_4th  ";
     }
   } else {
-    for (int n=0; n<outvars.size(); n++) {
+    for (size_t n=0; n<outvars.size(); n++) {
       msg << outvars[n].label.c_str() << "  ";
     }
   }
